pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flyxiv-observer"
version = "1.0.0"
description = "Discord bot for FFXIV event reminders and info scraping"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "FlyXIV Observer", email = "ns090200@gmail.com"}
]
keywords = ["discord", "bot", "ffxiv", "events", "reminders"]

dependencies = [
    "discord.py>=2.5.2",
    "python-dotenv>=1.0.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "beautifulsoup4>=4.13.4",
    "google-generativeai>=0.8.5"
]

[project.scripts]
flyxiv-observer = "pyobserver.main:main"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto" 